RF: Move attrib helper from resources.base to utils
We'll use this across the codebase, not just modules under resources/,
to avoid repeating attr.ib(..., default=None) many times.

This is pure code movement.
